@@ -1,47 +1,56 @@
-import React, { Component } from "react";
-import { Card, Avatar, Descriptions, notification, Divider, Tag} from "antd";
-import { UserOutlined } from "@ant-design/icons";
-import "../css/profile.css";
-import { connect } from "react-redux";
-import {getUserProfile} from '../modules/reducers/profileEffects'
+import React, { Component } from 'react';
+import {
+  Card, Avatar, Descriptions, notification, Divider, Tag,
+} from 'antd';
+import { UserOutlined } from '@ant-design/icons';
+import '../css/profile.css';
+import { connect } from 'react-redux';
+import PropTypes from 'prop-types';
+import { getUserProfile } from '../modules/reducers/profileEffects';
 
 class UserProfile extends Component {
-  componentDidMount(){
-    this.props.onPageLoad()
-    if (this.props.profile.setProfile === false) {
-      notification['info']({
-        message: `Your profile is loading......`,
-        description:this.props.profile.response,
+  componentDidMount() {
+    const {
+      profile, onPageLoad,
+    } = this.props;
+    onPageLoad();
+    if (profile.setProfile === false) {
+      notification.info({
+        message: 'Your profile is loading......',
+        description: profile.response,
         duration: 6,
-        placement:"topLeft"
+        placement: 'topLeft',
       });
-      }
+    }
   }
-  
+
   render() {
-    let isEmpty = "N/A"
-    if (this.props.profile.setProfile === true) {
-      notification['success']({
-        message: `Your profile is ready ${this.props.user.currentUser.username}`,
-        description:this.props.profile.response,
+    const {
+      profile, user, history,
+    } = this.props;
+    const isEmpty = 'N/A';
+    if (profile.setProfile === true) {
+      notification.success({
+        message: `Your profile is ready ${user.currentUser.username}`,
+        description: profile.response,
         duration: 2,
-        placement:"topLeft"
+        placement: 'topLeft',
       });
     }
     if (
-      this.props.user.isLogged === undefined ||
-      this.props.user.isLogged === false
+      user.isLogged === undefined
+      || user.isLogged === false
     ) {
-      notification["warning"]({
+      notification.warning({
         message:
-          "System resume failed, if it doesnt resume in a few, try to login again",
-        description: this.props.user.response,
+          'System resume failed, if it doesnt resume in a few, try to login again',
+        description: user.response,
         duration: 10,
-        placement: "bottomRight",
+        placement: 'bottomRight',
       });
-      this.props.history.push("/auth");
+      history.push('/auth');
     }
-    
+
     return (
       <div>
         <Card title="User profile">
@@ -50,28 +59,31 @@ class UserProfile extends Component {
               <Card bordered={false} className="avatar-card">
                 <Avatar shape="square" size={64} icon={<UserOutlined />} />
                 <Divider orientation="left">Username</Divider>
-                <Tag color="red">{this.props.user.currentUser.username || this.props.profile.user.username} </Tag>
+                <Tag color="red">
+                  {user.currentUser.username || profile.user.username}
+                  {' '}
+                </Tag>
               </Card>
             </div>
             <div>
               <Card type="inner" title="User profile">
                 <Descriptions layout="vertical">
                   <Descriptions.Item label="First name">
-                   <p>{this.props.profile.profile.first_name || isEmpty}</p> 
+                    <p>{profile.profile.first_name || isEmpty}</p>
                   </Descriptions.Item>
                   <Descriptions.Item label="Second name">
-                  {this.props.profile.profile.second_name || isEmpty }
+                    {profile.profile.second_name || isEmpty }
                   </Descriptions.Item>
                   <Descriptions.Item label="Telephone">
-                  {this.props.profile.profile.tel_number || isEmpty}
+                    {profile.profile.tel_number || isEmpty}
                   </Descriptions.Item>
                   <Descriptions.Item label="Location">
-                  {this.props.profile.profile.location || isEmpty}
+                    {profile.profile.location || isEmpty}
                   </Descriptions.Item>
                 </Descriptions>
               </Card>
               <Card type="inner" title="Acount type">
-    {this.props.user.currentUser.user_type}
+                {user.currentUser.user_type}
               </Card>
             </div>
           </div>
@@ -81,21 +93,58 @@ class UserProfile extends Component {
   }
 }
 
-const mapStateToProps = (state) => {
-  return {
-    user: state.user,
-    profile:state.profile
-  };
+const mapStateToProps = state => ({
+  user: state.user,
+  profile: state.profile,
+});
+
+const mapDispatchToProps = dispatch => ({
+  onPageLoad: () => {
+    dispatch(getUserProfile());
+  },
+});
+UserProfile.propTypes = {
+  profile: PropTypes.shape({
+    setProfile: PropTypes.bool,
+    response: PropTypes.string,
+    user: PropTypes.shape({
+      username: PropTypes.string,
+    }),
+    profile: PropTypes.shape({
+      first_name: PropTypes.string,
+      second_name: PropTypes.string,
+      tel_number: PropTypes.string,
+      location: PropTypes.string,
+
+    }),
+  }),
+  onPageLoad: PropTypes.func,
+  history: PropTypes.string,
+  user: PropTypes.shape({
+    currentUser: PropTypes.shape({
+      username: PropTypes.string,
+      user_type: PropTypes.string,
+    }),
+    isLogged: PropTypes.bool,
+    response: PropTypes.string,
+
+  }),
 };
+UserProfile.defaultProps = {
+  profile: PropTypes.shape({
+    saveProfile: false,
+    response: '',
+  }),
+  onPageLoad: () => {},
+  history: '/auth',
+  user: PropTypes.shape({
+    currentUser: PropTypes.shape({
+      username: '',
+      user_type: '',
+    }),
+    isLogged: false,
+    response: '',
 
-
-const mapDispatchToProps = (dispatch) => {
-  return {
-    onPageLoad: () => {
-      dispatch(getUserProfile());
-    }
-  };
+  }),
 };
-
-
 export default connect(mapStateToProps, mapDispatchToProps)(UserProfile);
