@@ -4,18 +4,17 @@ export const LOGIN_USER_ERROR = 'LOGIN_USER_ERROR';
 export const MAP_USER_SUCCESS = 'MAP_USER_SUCCESS';
 export const MAP_USER_ERROR = 'MAP_USER_ERROR';
 
-
-export const LOGOUT_USER_SUCCESS = 'LOGOUT_USER_SUCCESS'
+export const LOGOUT_USER_SUCCESS = 'LOGOUT_USER_SUCCESS';
 export const LOGOUT_USER_ERROR = 'LOGOUT_USER_ERROR';
 
-export const PROFILE_SUCCESS = "PROFILE_SUCCESS"
-export const PROFILE_ERROR = "PROFILE_ERROR"
+export const PROFILE_SUCCESS = 'PROFILE_SUCCESS';
+export const PROFILE_ERROR = 'PROFILE_ERROR';
 
-export const VET_SUCCESS = "VET_SUCCESS"
-export const VET_ERROR = "VET_ERROR"
+export const VET_SUCCESS = 'VET_SUCCESS';
+export const VET_ERROR = 'VET_ERROR';
 
-export const BOOKING_SUCCESS = "BOOKING_SUCCESS"
-export const BOOKING_ERROR = "BOOKING_ERROR"
+export const BOOKING_SUCCESS = 'BOOKING_SUCCESS';
+export const BOOKING_ERROR = 'BOOKING_ERROR';
 
-export const GET_BOOKING_SUCCESS = "GET_BOOKING_SUCCESS"
-export const GET_BOOKING_ERROR = "GET_BOOKING_ERROR"
+export const GET_BOOKING_SUCCESS = 'GET_BOOKING_SUCCESS';
+export const GET_BOOKING_ERROR = 'GET_BOOKING_ERROR';
