@@ -1,131 +1,105 @@
-import {loginUserSuccess, loginUserError, getUserSuccess, getUserError, logoutUserSuccess} from '../actions/actions'
 import axios from 'axios';
+import {
+  loginUserSuccess, loginUserError, getUserSuccess, getUserError, logoutUserSuccess,
+} from '../actions/actions';
 
-export const loginUser = ({ username, password }) => {
-  return (dispatch, getState) => {
-    axios
-      .post(`http://localhost:3000/api/v1/login`, {
-        user:{
-          username:username,
-          password:password
-        }
-      })
-      .then(res => {
-        if (res.data.status === "SUCCESS"){
-          setTimeout(() => {
-            localStorage.setItem('vet_token', res.data.jwt);
-        dispatch(loginUserSuccess(res.data));
-        }, 2500)
-
-        }
-        if (res.data.status === "FAIL"){
+export const loginUser = ({ username, password }) => dispatch => {
+  axios
+    .post('http://localhost:3000/api/v1/login', {
+      user: {
+        username,
+        password,
+      },
+    })
+    .then(res => {
+      if (res.data.status === 'SUCCESS') {
         setTimeout(() => {
-        dispatch(loginUserError(res.data));
-        }, 2500)
-        }
-        console.log('current state:', getState());
+          localStorage.setItem('vet_token', res.data.jwt);
+          dispatch(loginUserSuccess(res.data));
+        }, 2500);
+      }
+      if (res.data.status === 'FAIL') {
+        setTimeout(() => {
+          dispatch(loginUserError(res.data));
+        }, 2500);
+      }
+    })
 
-      })
-      
-      .catch(err => {
-        dispatch(loginUserError(err.message));
-        
-      });
-      
-  };
+    .catch(err => {
+      dispatch(loginUserError(err.message));
+    });
 };
 
-export const signupUser = ({email, nickname, password, confirm, user_type }) => {
-  return (dispatch, getState) => {
-
-    axios
-      .post(`http://localhost:3000/api/v1/user`, {
-        user:{
-          username:nickname,
-          password:password,
-          email:email,
-          password_confirmation:confirm,
-          user_type:user_type[0]
-        }
-      })
-      .then(res => {
-        if (res.data.status === "SUCCESS"){
-          setTimeout(() => {
-            localStorage.setItem('vet_token', res.data.jwt);
-        dispatch(loginUserSuccess(res.data));
-        }, 2500)
-
-        }
-        if (res.data.status === "FAIL"){
+export const signupUser = ({
+  // eslint-disable-next-line camelcase
+  email, nickname, password, confirm, user_type,
+}) => dispatch => {
+  axios
+    .post('http://localhost:3000/api/v1/user', {
+      user: {
+        username: nickname,
+        password,
+        email,
+        password_confirmation: confirm,
+        user_type: user_type[0],
+      },
+    })
+    .then(res => {
+      if (res.data.status === 'SUCCESS') {
         setTimeout(() => {
-        dispatch(loginUserError(res.data));
-        }, 2500)
-        }
-
-      })
-      .catch(err => {
-        dispatch(loginUserError(err));
-        
-      });
-      console.log('current err data:', getState())
-      
-  };
-};
-export const getUser = (token) => {
-  return (dispatch, getState) => {
-
-    axios
-      .get(`http://localhost:3000/api/v1/login/`, {
-        headers: {
-          'Authorization': `Basic ${token}`
-        }
-      })
-      .then(res => {
-        if (res.data.status === "SUCCESS"){
-          setTimeout(() => {
-          
-        dispatch(getUserSuccess(res.data));
-        }, 2500)
-
-        }
-        if (res.data.status === "FAIL"){
+          localStorage.setItem('vet_token', res.data.jwt);
+          dispatch(loginUserSuccess(res.data));
+        }, 2500);
+      }
+      if (res.data.status === 'FAIL') {
         setTimeout(() => {
-        dispatch(getUserError(res.data));
-        }, 2500)
-        }
-
-      })
-      .catch(err => {
-        dispatch(getUserError(err));
-        
-      });
-      console.log('current err data:', getState())
-      
-  };
+          dispatch(loginUserError(res.data));
+        }, 2500);
+      }
+    })
+    .catch(err => {
+      dispatch(loginUserError(err));
+    });
 };
-export const logOutUser = () => {
-  
-  return (dispatch, getState) => {
-    let getToken = localStorage.getItem("vet_token");
-    axios.get(`http://localhost:3000/api/v1/logout/`, {
-    
-        headers: {
-          'Authorization': `Basic ${getToken}`
-        }
-      })
-      .then(res => {
-        if (res.data.status === "SUCCESS"){
-          setTimeout(() => {
-        dispatch(logoutUserSuccess(res.data));
-        }, 2500)
+export const getUser = token => dispatch => {
+  axios
+    .get('http://localhost:3000/api/v1/login/', {
+      headers: {
+        Authorization: `Basic ${token}`,
+      },
+    })
+    .then(res => {
+      if (res.data.status === 'SUCCESS') {
+        setTimeout(() => {
+          dispatch(getUserSuccess(res.data));
+        }, 2500);
+      }
+      if (res.data.status === 'FAIL') {
+        setTimeout(() => {
+          dispatch(getUserError(res.data));
+        }, 2500);
+      }
+    })
+    .catch(err => {
+      dispatch(getUserError(err));
+    });
+};
+export const logOutUser = () => dispatch => {
+  const getToken = localStorage.getItem('vet_token');
+  axios.get('http://localhost:3000/api/v1/logout/', {
 
-        }
-
-      })
-      .catch(err => {
-        dispatch(logoutUserSuccess(err));
-        
-      });
-      
-  };
+    headers: {
+      Authorization: `Basic ${getToken}`,
+    },
+  })
+    .then(res => {
+      if (res.data.status === 'SUCCESS') {
+        setTimeout(() => {
+          dispatch(logoutUserSuccess(res.data));
+        }, 2500);
+      }
+    })
+    .catch(err => {
+      dispatch(logoutUserSuccess(err));
+    });
 };
