@@ -1,62 +1,60 @@
-import { VetSuccess, VetError, BookSuccess, BookError } from "../actions/actions";
-import axios from "axios";
+import axios from 'axios';
+import {
+  VetSuccess, VetError, BookSuccess, BookError,
+} from '../actions/actions';
 
-export const getVetsProfile = () => {
-  return (dispatch) => {
-    let getToken = localStorage.getItem("vet_token");
+export const getVetsProfile = () => dispatch => {
+  const getToken = localStorage.getItem('vet_token');
 
-    axios
-      .get(`http://localhost:3000/api/v1/appointment/`, {
-        headers: {
-          Authorization: `Basic ${getToken}`,
-        },
-      })
-      .then((res) => {
-        if (res.data.status === "SUCCESS") {
-          setTimeout(() => {
-            dispatch(VetSuccess(res.data));
-          }, 2500);
-        }
-        if (res.data.status === "FAIL") {
-          setTimeout(() => {
-            dispatch(VetError(res.data));
-          }, 2500);
-        }
-      })
-      .catch((err) => {
-        dispatch(VetError(err));
-      });
-  };
+  axios
+    .get('http://localhost:3000/api/v1/appointment/', {
+      headers: {
+        Authorization: `Basic ${getToken}`,
+      },
+    })
+    .then(res => {
+      if (res.data.status === 'SUCCESS') {
+        setTimeout(() => {
+          dispatch(VetSuccess(res.data));
+        }, 2500);
+      }
+      if (res.data.status === 'FAIL') {
+        setTimeout(() => {
+          dispatch(VetError(res.data));
+        }, 2500);
+      }
+    })
+    .catch(err => {
+      dispatch(VetError(err));
+    });
 };
 
-export const postBookingProfile = (data) => {
-  return (dispatch) => {
-    let getToken = localStorage.getItem("vet_token");
-   
-    const headers = {
-      'Content-Type': 'application/json',
-      'Authorization': `Basic ${getToken}`
-    }
+export const postBookingProfile = data => dispatch => {
+  const getToken = localStorage.getItem('vet_token');
 
-    axios
-      .post(`http://localhost:3000/api/v1/appointment/`, data, {
-        "headers": headers
- 
-      })
-      .then((res) => {
-        if (res.data.status === "SUCCESS") {
-          setTimeout(() => {
-            dispatch(BookSuccess(res.data));
-          }, 2500);
-        }
-        if (res.data.status === "FAIL") {
-          setTimeout(() => {
-            dispatch(BookError(res.data));
-          }, 2500);
-        }
-      })
-      .catch((err) => {
-        dispatch(BookError(err));
-      });
+  const headers = {
+    'Content-Type': 'application/json',
+    Authorization: `Basic ${getToken}`,
   };
+
+  axios
+    .post('http://localhost:3000/api/v1/appointment/', data, {
+      headers,
+
+    })
+    .then(res => {
+      if (res.data.status === 'SUCCESS') {
+        setTimeout(() => {
+          dispatch(BookSuccess(res.data));
+        }, 2500);
+      }
+      if (res.data.status === 'FAIL') {
+        setTimeout(() => {
+          dispatch(BookError(res.data));
+        }, 2500);
+      }
+    })
+    .catch(err => {
+      dispatch(BookError(err));
+    });
 };
