@@ -1,8 +1,9 @@
-import { GET_BOOKING_SUCCESS, GET_BOOKING_ERROR } from "../actions/actionTypes";
+import { GET_BOOKING_SUCCESS, GET_BOOKING_ERROR } from '../actions/actionTypes';
+
 const INITIAL_STATE = {
   Appointments: {},
-  response: "",
-  loading: false
+  response: '',
+  loading: false,
 };
 
 const myBookings = (state = INITIAL_STATE, action) => {
@@ -12,14 +13,14 @@ const myBookings = (state = INITIAL_STATE, action) => {
         ...state,
         Appointments: action.payload.profile,
         response: action.payload.messages,
-        loading:true
+        loading: true,
       };
     case GET_BOOKING_ERROR:
       return {
         ...state,
         Appointments: {},
         response: action.payload.errors,
-        loading:false
+        loading: false,
       };
     default:
       return state;
